a safe version of rm
When I rm something and it was a mistake, I usually notice right away.
Therefore, instead of replacing rm with a move-to-trash script which uses
up disk space forever, it is better to periodically remove the files I
have least-recently removed.

I called the script "un" (for uncreate) instead of "rm" because I don't
want to start typing "rm" with confidence, knowing the files are backed up
somewhere, only to discover that my script wasn't installed on that
machine.