Excel to TXT
Now that the Excel format is a zip of XML files instead of proprietary
binary blob, it is possible (but not straightforward) to convert some
of it into plain tab-separated text files. Of course, Excel documents
contain a lot more than just text, so the main usecase is for those
times when people send you an Excel because it's the simplest way they
know of to maintain tabular data.