Bash-based delta-debugging
Delta debugging (http://www.st.cs.uni-saarland.de/dd/) is an awesome way
to debug: write a script to decide whether the bug is present, and let the
machine search through all the combinations and find you a minimum sample
case.

The original code is in python, but I find it easier to write such scripts
in bash, so I wrote a simple wrapper. Usage:

  > deltad test.sh < test.in