Visual Studio to Makefile
If your Visual Studio project is simple enough and cross-platform
enough that the same code could be run from a simple Makefile...
Then it can! This script will convert part of a Visual Studio
solution file to a Makefile script. Advanced (and not that advanced)
Visual Studio features are ignored.