git-diff-interactive
Interactively modifies the work tree towards the given commit.

I have struggled to pick an appropriate name for this; it's not quite
like "git-diff COMMIT", because the diff is in the opposite direction.
It's a bit like "git checkout COMMIT -- ." and "git reset COMMIT -- .",
but it's nothing like "git checkout COMMIT" nor "git reset COMMIT". It's
certainly not a pull, or a merge, either. Maybe apply?