diff3 version of diffconflicts
see https://github.com/whiteinge/dotfiles/blob/master/bin/diffconflicts

diffconflicts has a 2-way vimdiff mode which only displays the conflicts,
and a 3-way vimdiff which shows all the differences. To resolve
conflicts, you must create a file which contains both the changes on the
left and the changes on the right. Except that without a common
ancestor, how can you tell if a function which is missing on one side
was added on the right or deleted on the left?

diff3conflicts is a 3-way vimdiff mode which only displays the
conflicts. To resolve conflicts, modify the middle pane so that it
incorporates the changes on the left and the changes on the right. If a
function is missing on one side and present in the middle, it was
deleted. If it is missing on one side and missing in the middle, it was
added.