fixed: "folder 40 already exists"
"un" cleans up its trash on a rolling basis, keeping the most
recent items, 40 by default. Once you delete more than the max
number of files (50 by default), the oldest 10 are be deleted
to keep the count down to 40. This is usually files from a few
days ago, but if you delete 50 files in a single day, it breaks
an internal invariant according to which files are stored in
folders corresponding to dates, in folders numbered from 0: if
there are 50 files today and you delete the 10 oldest ones, then
the 40 folders for today will start counting at 11. This fix
changes the logic so that we no longer rely on this invariant.