different bin files for nix and fedora
I'll probably need a different one for macOS as well.

The purpose of this repo is to allow me to easily configure a new
machine with the customizations and tools I've grown acustomed to over
the years, but I've struggled to keep those scripts portable. When I
switch to a new machine, I use this repo to install all the scripts
as-is, and as time goes by, I tweak them in ways which might only work
on that machine without realizing it. I'll only notice on the next
machine, at which point I'll "fix" the script, probably in a way which
only works on that new machine.

This approach is no longer workable, because my latest machine
dual-boots Fedora and NixOS. So I need to make the scripts work with
both. My plan is to keep the current scripts (which have been tweaked
for my previous machine, a MacOS laptop) in bin, and then to slowly
migrate them to bin-common if I can make them portable enough so they
work on both Fedora and NixOS, or to bin-fedora or bin-nix if I can't.