give 'bin' a less conspicuous name
'bin' was the folder in which I accumulated small utility scripts over
the years. I now need to go over each of them and make them
nix-compatible (which is usually as simple as using "#!/usr/bin/env
bash" instead of "#!/bin/bash", but still). Going over the list makes me
realize that I haven't used some of those scripts in years, so I think
I'll move the remaining scripts out of the way, and only nixify and
bring back into the PATH those scripts I actually do use. That is, if in
the future I find myself typing a command and it's not there, I'll
nixify it then.